+++ lib/service/service.dart
@@ -1,28 +1,9 @@
 import 'package:amiibo_network/dao/SQLite/amiibo_sqlite.dart';
-import 'package:flutter/foundation.dart';
 import '../model/amiibo_local_db.dart';
 import 'dart:convert';
-import 'package:flutter/services.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-
-final AmiiboSQLite dao = AmiiboSQLite();
-
-Future<void> initDB() async => await dao.initDB();
 
 class Service{
-  static Map<String, dynamic> _jsonFile;
-  static DateTime _lastUpdate;
-  static DateTime _lastUpdateDB;
-
-  static final Service _instance = Service._();
-  factory Service() => _instance;
-  Service._();
-
-  Future<Map<String, dynamic>> get jsonFile async {
-    return _jsonFile ??= jsonDecode(await rootBundle.loadString('assets/databases/amiibos.json'));
-  }
-
-  Future<AmiiboLocalDB> fetchAllAmiibo() async => compute(entityFromMap, await jsonFile);
+  final AmiiboSQLite dao = AmiiboSQLite();
 
   Future<AmiiboLocalDB> fetchAllAmiiboDB([String orderBy]) => dao.fetchAll(orderBy);
 
@@ -37,34 +18,6 @@ class Service{
     return jsonEncode(amiibos);
   }
 
-  Future<DateTime> get lastUpdateDB async{
-    final SharedPreferences preferences = await SharedPreferences.getInstance();
-    return _lastUpdateDB ??= DateTime.tryParse(preferences.getString('Date') ?? '');
-  }
-
-  Future<DateTime> get lastUpdate async{
-    return _lastUpdate ??= DateTime.tryParse((await jsonFile)['lastUpdated'] ?? '');
-  }
-
-  Future<bool> createDB() async{
-    return compareLastUpdate().then((sameDate) async {
-      if(sameDate == null) throw Exception("Couldn't fetch last update");
-      if(!sameDate) fetchAllAmiibo().then(_updateDB);
-      return await Future.value(true);
-    }).catchError((e) {
-      print(e.toString());
-      return false;
-    });
-  }
-
-  _updateDB(AmiiboLocalDB amiiboDB) async{
-    dao.insertAll(amiiboDB, "amiibo").then((_) async{
-      final SharedPreferences preferences = await SharedPreferences.getInstance();
-      final DateTime dateTime = await lastUpdate;
-      await preferences.setString('Date', dateTime?.toIso8601String());
-    });
-  }
-
   Future<AmiiboLocalDB> fetchByCategory([String column, List<String> args,
     String orderBy]) {
     if(column == null || (args?.isEmpty ?? true)) column = args = null;
@@ -73,23 +26,13 @@ class Service{
 
   Future<AmiiboDB> fetchAmiiboDBByKey(String key) => dao.fetchByKey(key);
 
-  Future<void> update(AmiiboLocalDB amiibos) {
-    return dao.insertImport(amiibos);
-  }
+  Future<void> update(AmiiboLocalDB amiibos) => dao.insertImport(amiibos);
 
   Future<List<String>> fetchDistinct([bool figure = true]) async =>
     dao.fetchDistinct('amiibo', 'amiiboSeries',
       figure ? 'type IS NOT' : 'type IS', ["Card"]);
 
-  Future<List<String>> searchDB(String name) async =>
-    dao.fetchLimit('%$name%', 10);
-
-  Future<bool> compareLastUpdate() async{
-    final dateDB = await lastUpdateDB;
-    final dateJson = await lastUpdate;
-
-    return dateDB?.isAtSameMomentAs(dateJson) ?? false;
-  }
+  Future<List<String>> searchDB(String name) async => dao.fetchLimit('%$name%', 10);
 
   Future<void> resetCollection() async{
     await dao.updateAll('amiibo', {'wishlist' : 0, 'owned' : 0});